+++ src/libnml/cms/tcp_srv.cc
@@ -434,9 +434,9 @@ void CMS_SERVER_REMOTE_TCP_PORT::run()
 			    blocking_thread_kill
 				(client_port_to_check->threadId);
 #if 0
-			    *((u_long *) temp_buffer) =
+			    *((uint32_t *) temp_buffer) =
 				htonl(client_port_to_check->serial_number);
-			    *((u_long *) temp_buffer + 1) =
+			    *((uint32_t *) temp_buffer + 1) =
 				htonl((unsigned long)
 				CMS_SERVER_SIDE_ERROR);
 			    putbe32(temp_buffer + 8, 0);	/* size
@@ -696,8 +696,8 @@ void CMS_SERVER_REMOTE_TCP_PORT::handle_request(CLIENT_TCP_PORT *
 	_client_tcp_port->errors++;
     }
     _client_tcp_port->serial_number++;
-    request_type = ntohl(*((u_long *) temp_buffer + 1));
-    buffer_number = ntohl(*((u_long *) temp_buffer + 2));
+    request_type = ntohl(*((uint32_t *) temp_buffer + 1));
+    buffer_number = ntohl(*((uint32_t *) temp_buffer + 2));
 
     rcs_print_debug(PRINT_ALL_SOCKET_REQUESTS,
 	"TCPSVR request recieved: fd = %d, serial_number=%ld, request_type=%ld, buffer_number=%ld\n",
@@ -758,9 +758,9 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 	    memcpy(_client_tcp_port->diag_info->host_sysinfo,
 		server->set_diag_info_buf + 16, 32);
 	    _client_tcp_port->diag_info->pid =
-		htonl(*((u_long *) (server->set_diag_info_buf + 48)));
+		htonl(*((uint32_t *) (server->set_diag_info_buf + 48)));
 	    _client_tcp_port->diag_info->c_num =
-		htonl(*((u_long *) (server->set_diag_info_buf + 52)));
+		htonl(*((uint32_t *) (server->set_diag_info_buf + 52)));
 	    memcpy(&(_client_tcp_port->diag_info->rcslib_ver),
 		server->set_diag_info_buf + 56, 8);
 	    _client_tcp_port->diag_info->reverse_flag =
@@ -828,7 +828,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 		    dpi_offset += 16;
 		    memcpy(temp_buffer + dpi_offset, dpi->host_sysinfo, 32);
 		    dpi_offset += 32;
-		    *((u_long *) ((char *) temp_buffer + dpi_offset)) =
+		    *((uint32_t *) ((char *) temp_buffer + dpi_offset)) =
 			htonl(dpi->pid);
 		    dpi_offset += 4;
 		    if (_client_tcp_port->diag_info->reverse_flag ==
@@ -842,22 +842,22 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 			    8);
 		    }
 		    dpi_offset += 8;
-		    *((u_long *) ((char *) temp_buffer + dpi_offset)) =
+		    *((uint32_t *) ((char *) temp_buffer + dpi_offset)) =
 			htonl(dpi->access_type);
 		    dpi_offset += 4;
-		    *((u_long *) ((char *) temp_buffer + dpi_offset)) =
+		    *((uint32_t *) ((char *) temp_buffer + dpi_offset)) =
 			htonl(dpi->msg_id);
 		    dpi_offset += 4;
-		    *((u_long *) ((char *) temp_buffer + dpi_offset)) =
+		    *((uint32_t *) ((char *) temp_buffer + dpi_offset)) =
 			htonl(dpi->msg_size);
 		    dpi_offset += 4;
-		    *((u_long *) ((char *) temp_buffer + dpi_offset)) =
+		    *((uint32_t *) ((char *) temp_buffer + dpi_offset)) =
 			htonl(dpi->msg_type);
 		    dpi_offset += 4;
-		    *((u_long *) ((char *) temp_buffer + dpi_offset)) =
+		    *((uint32_t *) ((char *) temp_buffer + dpi_offset)) =
 			htonl(dpi->number_of_accesses);
 		    dpi_offset += 4;
-		    *((u_long *) ((char *) temp_buffer + dpi_offset)) =
+		    *((uint32_t *) ((char *) temp_buffer + dpi_offset)) =
 			htonl(dpi->number_of_new_messages);
 		    dpi_offset += 4;
 		    if (_client_tcp_port->diag_info->reverse_flag ==
@@ -928,12 +928,12 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 		    dpi_offset += 8;
 		    int is_last_writer =
 			(dpi == diagreply->cdi->last_writer_dpi);
-		    *((u_long *) ((char *) temp_buffer + dpi_offset)) =
+		    *((uint32_t *) ((char *) temp_buffer + dpi_offset)) =
 			htonl(is_last_writer);
 		    dpi_offset += 4;
 		    int is_last_reader =
 			(dpi == diagreply->cdi->last_reader_dpi);
-		    *((u_long *) ((char *) temp_buffer + dpi_offset)) =
+		    *((uint32_t *) ((char *) temp_buffer + dpi_offset)) =
 			htonl(is_last_reader);
 		    dpi_offset += 4;
 		    dpi =
@@ -941,8 +941,8 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 			get_next();
 		}
 	    }
-	    *((u_long *) temp_buffer + 6) = htonl(dpi_count);
-	    *((u_long *) temp_buffer + 7) = htonl(dpi_offset);
+	    *((uint32_t *) temp_buffer + 6) = htonl(dpi_count);
+	    *((uint32_t *) temp_buffer + 7) = htonl(dpi_offset);
 	    if (sendn
 		(_client_tcp_port->socket_fd, temp_buffer, dpi_offset, 0,
 		    dtimeout) < 0) {
@@ -1000,9 +1000,9 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 #endif
 	    blocking_read_req->buffer_number = buffer_number;
 	    blocking_read_req->access_type =
-		ntohl(*((u_long *) temp_buffer + 3));
+		ntohl(*((uint32_t *) temp_buffer + 3));
 	    blocking_read_req->last_id_read =
-		ntohl(*((u_long *) temp_buffer + 4));
+		ntohl(*((uint32_t *) temp_buffer + 4));
 	    total_subdivisions = 1;
 	    if (max_total_subdivisions > 1) {
 		total_subdivisions =
@@ -1011,7 +1011,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 	    if (total_subdivisions > 1) {
 		if (recvn
 		    (_client_tcp_port->socket_fd,
-			(char *) (((u_long *) temp_buffer) + 5), 8, 0, -1,
+			(char *) (((uint32_t *) temp_buffer) + 5), 8, 0, -1,
 			NULL) < 0) {
 		    rcs_print_error
 			("Can not read from client port (%d) from %s\n",
@@ -1021,11 +1021,11 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 		    return;
 		}
 		blocking_read_req->subdiv =
-		    ntohl(*((u_long *) temp_buffer + 6));
+		    ntohl(*((uint32_t *) temp_buffer + 6));
 	    } else {
 		if (recvn
 		    (_client_tcp_port->socket_fd,
-			(char *) (((u_long *) temp_buffer) + 5), 4, 0, -1,
+			(char *) (((uint32_t *) temp_buffer) + 5), 4, 0, -1,
 			NULL) < 0) {
 		    rcs_print_error
 			("Can not read from client port (%d) from %s\n",
@@ -1036,7 +1036,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 		}
 	    }
 	    blocking_read_req->timeout_millis =
-		ntohl(*((u_long *) temp_buffer + 5));
+		ntohl(*((uint32_t *) temp_buffer + 5));
 	    blocking_read_req->server = server;
 	    blocking_read_req->remport = this;
 	    _client_tcp_port->blocking = 1;
@@ -1053,9 +1053,9 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 		    thr_retval);
 		rcs_print_error("pthread_create error: %d %s\n", errno,
 		    strerror(errno));
-		*((u_long *) temp_buffer) =
+		*((uint32_t *) temp_buffer) =
 		    htonl(_client_tcp_port->serial_number);
-		*((u_long *) temp_buffer + 1) =
+		*((uint32_t *) temp_buffer + 1) =
 		    htonl((unsigned long) CMS_SERVER_SIDE_ERROR);
 		putbe32(temp_buffer + 8, 0);	/* size */
 		putbe32(temp_buffer + 12, 0);	/* write_id */
@@ -1094,9 +1094,9 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 #else
 	    rcs_print_error
 		("Blocking read not supported on this platform.\n");
-	    *((u_long *) temp_buffer) =
+	    *((uint32_t *) temp_buffer) =
 		htonl(_client_tcp_port->serial_number);
-	    *((u_long *) temp_buffer + 1) =
+	    *((uint32_t *) temp_buffer + 1) =
 		htonl((unsigned long) CMS_SERVER_SIDE_ERROR);
 	    putbe32(temp_buffer + 8, 0);	/* size */
 	    putbe32(temp_buffer + 12, 0);	/* write_id */
@@ -1112,8 +1112,8 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 
     case REMOTE_CMS_READ_REQUEST_TYPE:
 	server->read_req.buffer_number = buffer_number;
-	server->read_req.access_type = ntohl(*((u_long *) temp_buffer + 3));
-	server->read_req.last_id_read = ntohl(*((u_long *) temp_buffer + 4));
+	server->read_req.access_type = ntohl(*((uint32_t *) temp_buffer + 3));
+	server->read_req.last_id_read = ntohl(*((uint32_t *) temp_buffer + 4));
 	server->read_reply =
 	    (REMOTE_READ_REPLY *) server->process_request(&server->read_req);
 	if (max_total_subdivisions > 1) {
@@ -1123,7 +1123,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 	if (total_subdivisions > 1) {
 	    if (recvn
 		(_client_tcp_port->socket_fd,
-		    (char *) (((u_long *) temp_buffer) + 5), 4, 0, -1,
+		    (char *) (((uint32_t *) temp_buffer) + 5), 4, 0, -1,
 		    NULL) < 0) {
 		rcs_print_error
 		    ("Can not read from client port (%d) from %s\n",
@@ -1132,7 +1132,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 		_client_tcp_port->errors++;
 		return;
 	    }
-	    server->read_req.subdiv = ntohl(*((u_long *) temp_buffer + 5));
+	    server->read_req.subdiv = ntohl(*((uint32_t *) temp_buffer + 5));
 	} else {
 	    server->read_req.subdiv = 0;
 	}
@@ -1181,8 +1181,8 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 
     case REMOTE_CMS_WRITE_REQUEST_TYPE:
 	server->write_req.buffer_number = buffer_number;
-	server->write_req.access_type = ntohl(*((u_long *) temp_buffer + 3));
-	server->write_req.size = ntohl(*((u_long *) temp_buffer + 4));
+	server->write_req.access_type = ntohl(*((uint32_t *) temp_buffer + 3));
+	server->write_req.size = ntohl(*((uint32_t *) temp_buffer + 4));
 	total_subdivisions = 1;
 	if (max_total_subdivisions > 1) {
 	    total_subdivisions =
@@ -1191,7 +1191,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 	if (total_subdivisions > 1) {
 	    if (recvn
 		(_client_tcp_port->socket_fd,
-		    (char *) (((u_long *) temp_buffer) + 5), 4, 0, -1,
+		    (char *) (((uint32_t *) temp_buffer) + 5), 4, 0, -1,
 		    NULL) < 0) {
 		rcs_print_error
 		    ("Can not read from client port (%d) from %s\n",
@@ -1200,7 +1200,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 		_client_tcp_port->errors++;
 		return;
 	    }
-	    server->write_req.subdiv = ntohl(*((u_long *) temp_buffer + 5));
+	    server->write_req.subdiv = ntohl(*((uint32_t *) temp_buffer + 5));
 	} else {
 	    server->write_req.subdiv = 0;
 	}
@@ -1243,7 +1243,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
     case REMOTE_CMS_CHECK_IF_READ_REQUEST_TYPE:
 	server->check_if_read_req.buffer_number = buffer_number;
 	server->check_if_read_req.subdiv =
-	    ntohl(*((u_long *) temp_buffer + 3));
+	    ntohl(*((uint32_t *) temp_buffer + 3));
 	server->check_if_read_reply =
 	    (REMOTE_CHECK_IF_READ_REPLY *) server->process_request(&server->
 	    check_if_read_req);
@@ -1256,9 +1256,9 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 	    return;
 	}
 	putbe32(temp_buffer, _client_tcp_port->serial_number);
-	*((u_long *) temp_buffer + 1) =
+	*((uint32_t *) temp_buffer + 1) =
 	    htonl(server->check_if_read_reply->status);
-	*((u_long *) temp_buffer + 2) =
+	*((uint32_t *) temp_buffer + 2) =
 	    htonl(server->check_if_read_reply->was_read);
 	if (sendn(_client_tcp_port->socket_fd, temp_buffer, 12, 0, dtimeout) <
 	    0) {
@@ -1269,7 +1269,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
     case REMOTE_CMS_GET_MSG_COUNT_REQUEST_TYPE:
 	server->get_msg_count_req.buffer_number = buffer_number;
 	server->get_msg_count_req.subdiv =
-	    ntohl(*((u_long *) temp_buffer + 3));
+	    ntohl(*((uint32_t *) temp_buffer + 3));
 	server->get_msg_count_reply =
 	    (REMOTE_GET_MSG_COUNT_REPLY *) server->process_request(&server->
 	    get_msg_count_req);
@@ -1282,9 +1282,9 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 	    return;
 	}
 	putbe32(temp_buffer, _client_tcp_port->serial_number);
-	*((u_long *) temp_buffer + 1) =
+	*((uint32_t *) temp_buffer + 1) =
 	    htonl(server->get_msg_count_reply->status);
-	*((u_long *) temp_buffer + 2) =
+	*((uint32_t *) temp_buffer + 2) =
 	    htonl(server->get_msg_count_reply->count);
 	if (sendn(_client_tcp_port->socket_fd, temp_buffer, 12, 0, dtimeout) <
 	    0) {
@@ -1295,7 +1295,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
     case REMOTE_CMS_GET_QUEUE_LENGTH_REQUEST_TYPE:
 	server->get_queue_length_req.buffer_number = buffer_number;
 	server->get_queue_length_req.subdiv =
-	    ntohl(*((u_long *) temp_buffer + 3));
+	    ntohl(*((uint32_t *) temp_buffer + 3));
 	server->get_queue_length_reply =
 	    (REMOTE_GET_QUEUE_LENGTH_REPLY *) server->
 	    process_request(&server->get_queue_length_req);
@@ -1308,9 +1308,9 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 	    return;
 	}
 	putbe32(temp_buffer, _client_tcp_port->serial_number);
-	*((u_long *) temp_buffer + 1) =
+	*((uint32_t *) temp_buffer + 1) =
 	    htonl(server->get_queue_length_reply->status);
-	*((u_long *) temp_buffer + 2) =
+	*((uint32_t *) temp_buffer + 2) =
 	    htonl(server->get_queue_length_reply->queue_length);
 	if (sendn(_client_tcp_port->socket_fd, temp_buffer, 12, 0, dtimeout) <
 	    0) {
@@ -1321,7 +1321,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
     case REMOTE_CMS_GET_SPACE_AVAILABLE_REQUEST_TYPE:
 	server->get_space_available_req.buffer_number = buffer_number;
 	server->get_space_available_req.subdiv =
-	    ntohl(*((u_long *) temp_buffer + 3));
+	    ntohl(*((uint32_t *) temp_buffer + 3));
 	server->get_space_available_reply =
 	    (REMOTE_GET_SPACE_AVAILABLE_REPLY *) server->
 	    process_request(&server->get_space_available_req);
@@ -1334,9 +1334,9 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 	    return;
 	}
 	putbe32(temp_buffer, _client_tcp_port->serial_number);
-	*((u_long *) temp_buffer + 1) =
+	*((uint32_t *) temp_buffer + 1) =
 	    htonl(server->get_space_available_reply->status);
-	*((u_long *) temp_buffer + 2) =
+	*((uint32_t *) temp_buffer + 2) =
 	    htonl(server->get_space_available_reply->space_available);
 	if (sendn(_client_tcp_port->socket_fd, temp_buffer, 12, 0, dtimeout) <
 	    0) {
@@ -1346,7 +1346,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 
     case REMOTE_CMS_CLEAR_REQUEST_TYPE:
 	server->clear_req.buffer_number = buffer_number;
-	server->clear_req.subdiv = ntohl(*((u_long *) temp_buffer + 3));
+	server->clear_req.subdiv = ntohl(*((uint32_t *) temp_buffer + 3));
 	server->clear_reply =
 	    (REMOTE_CLEAR_REPLY *) server->process_request(&server->
 	    clear_req);
@@ -1454,9 +1454,9 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
     case REMOTE_CMS_SET_SUBSCRIPTION_REQUEST_TYPE:
 	server->set_subscription_req.buffer_number = buffer_number;
 	server->set_subscription_req.subscription_type =
-	    ntohl(*((u_long *) temp_buffer + 3));
+	    ntohl(*((uint32_t *) temp_buffer + 3));
 	server->set_subscription_req.poll_interval_millis =
-	    ntohl(*((u_long *) temp_buffer + 4));
+	    ntohl(*((uint32_t *) temp_buffer + 4));
 	server->set_subscription_reply =
 	    (REMOTE_SET_SUBSCRIPTION_REPLY *) server->
 	    process_request(&server->set_subscription_req);
@@ -1485,7 +1485,7 @@ void CMS_SERVER_REMOTE_TCP_PORT::switch_function(CLIENT_TCP_PORT *
 		}
 	    }
 	    putbe32(temp_buffer, _client_tcp_port->serial_number);
-	    *((u_long *) temp_buffer + 1) =
+	    *((uint32_t *) temp_buffer + 1) =
 		htonl(server->set_subscription_reply->success);
 	    /* successful ? */
 	    sendn(_client_tcp_port->socket_fd, temp_buffer, 8, 0, dtimeout);
